Separated order card of salesman and driver
